packet macro supports variable sized packets
This implementation is incredibly brittle right now, in particular the
part that generates code to calculate the size that a packet will have.

Ideally we would serialize the packet and count the bytes from that,
thus automatically relying upon the existing serialization logic.

That's not really what serde wants us to do however, because it supports
many different approaches to serialization, not just binary streams.

So for the moment we do some hacky approach that special cases
particular types that we know will serialize to a variable length (just
Vec at the moment).