Serialization uses actual Packet struct
This significantly cleans up the packet macro code, as we're no longer
generating individual packet (and packet impl) structs for each use of
the macro. This also allows us to remove the unsafe code that it was
using.

Now in order to serialize a packet correctly it is necessary to call
to_packet on the structure with the packet contents, and serialize the
resulting Packet struct.