Serializer tracks size of serialized data
Because the size calculations need to be kept synchronized with any
changes to the actual writer output, I thought it best to keep the logic
in the same Serializer type rather than create a separate one just for
calculating the size.

The writer field has also been made optional, so that it is possible to
construct a Serializer which only calculates the size without actually
writing the serialized output anywhere.

We need to write a solid test suite for the Serializer to ensure that
the sizes calculated for each of the serialization routines is accurate.