Packet size is calculated with Serializer
We now calculate the size for variable-sized packets using the
Serializer's own ability to calculate the correct size for its
serialization output.

This requires recursing through the entire packet structure each time it
gets serialized, which is more work at runtime than the previous
approach (which tried to calculate as much as possible statically).

However, this way is significantly more reliable and just overall more
sane.