Basic implementation of login server for testing
All of this structure needs to be re-designed with decent interfaces
between functional layers, an explicit state machine model for the login
process, and support for multiple clients through concurrency and
probably async i/o and events.

But for now it works to test that the ClassicUO client can actually move
through the login process until hand-off to the game (or relay) server.