Macro for defining codecs for sets of UO packets
Each codec can be used with tokio_util::codec::Framed to create a Sink
and/or Stream for sending and receiving the specified UO packet types.

In general it would be possible to define a single codec that included
all packets types in the protocol. However, to get the most value out of
type safety we will model the UO protocol as a state machine, where each
state has an associated codec that only works with the packet types that
are valid to send/receive in that state.

As such this macro helps cut down on the boilerplate code for defining
each codec.