List type with defined length prefix size
This is necessary for the UO protocol because sometimes it has sequences
that are prefixed with an 8-bit value, and othertimes with a 16-bit
value. That's what I've encountered so far at least, they might also use
32-bit and 64-bit length prefixes in some cases, so I've got support for
those as well.

This includes the serialization and deserialization logic, but it does
not yet compile.