Fixes to Huffman coding compression algo
When calculating the portion of a symbol's bit sequence that would need
to be OR'd onto the current byte we needed to be masking off the other
bits to satisfy Rust. However, we were accidentally masking off all the
other bits instead, so we needed to invert the mask.

Also, we were not correctly handling cases where the length of a bit
sequence was longer than the remaining bits for the current byte AND the
next byte (e.g. an 11 bit sequence when you have 1 or 2 bits remaining
in the current byte).