Redesign the define_codec! macro output
In particular, we only use a frame enum for the Decoder implementation.

For the Encoder side we use a <codec name>Encode trait that is impl'd by
each packet type which can be sent with the codec. And an implementation
of Encoder is generated that is generic over its item type which is
constrainted to have the appropriate Encode trait.