Login server is async & supports multiple clients
This makes use of the FSM and codecs defined in
uoverse_server::login::client, and uses tokio for the async runtime.

The FSM and codecs have been significantly reworked from their previous
design which had the state types containing the login server logic
directly.

In the new model each state contains the functionality that is unique to
that state, such as the ability to send or receive packets according to
a particular codec. The login logic remains implemented separately, in
the login binary.