Serialization type for null-terminated lists
In addition to lists that have a length prefix, the UO network protocol
also makes use of lists which are terminated with a sentinel
value (always zero from what I've seen so far).

Previously we were using the default `Vec` serde code, but it is better
to have an explicit type for these cases. Additionally this means that
the sequence de/serialization code is more generic, because it gets
reused by both `List` and now `ListTerm`.

NOTE: I'm pretty sure that the deserialization code for both of these
list types is not working properly right now, but I haven't taken the
time to test and fix it yet.