Client state machines retain framed i/o buffer
Previously we were constructing a new `FramedRead` or `FramedWrite` each
time the `recv()` or `send()` methods were called on a given state.

The problem with this approach is that the framed i/o types hold a
buffer internally, and when we dropped the old state we lost any
remaining data in that buffer.

This was only an issue for us with `FramedRead` where we might have read
packet data that should have been available to the next state, but was
instead thrown away.

The new approach constructs a `Framed` (which we should have been using
anyways since it works as both read/write types) when the client state
is constructed, using `map_codec` to build it using parts from the
previous `Framed` but with the codec changed to whatever is needed for
the new state.