Client states split between pre-world & in-world
The pre-world states constitute the game server side of the login
process, including the handshake upon coming from the login server as
well as character creation and/or selection.

The in-world states cover everything that takes place after a character
has been selected and entered the world. This state involves interacting
with the Server simulating the world.