Various early in-world packets from client
These were observed being sent from the ClassicUO client when it
connected to the server.

In order to prevent the server from disconnecting the client, we need to
handle every kind of packet we receive. Since these are sent
automatically upon entering the world, we at least need to accept them
otherwise the client will immediately disconnect.