UO protocol List type deserializes properly
The previous approach could not work because we were attempting to
deserialize the elements as a Vec, which has a Deserialize
implementation that maps to a sequence in serde's data model, which
doesn't have a defined length.

Instead, we have a new type (ListElements) that wraps a Vec and includes
the expected length. We implement DeserializeSeed for it, mapping to
serde's tuple model type.

This allows us to deserialize the length as the first element in a
struct, construct the ListElements with that as the expected type, then
attempt to deserialize a Vec using that as the seed.

This likely still does not support other Deserializer implementations,
if their struct deserialization uses something other than a
sequence (e.g. a map).