List type uses length type parameter
Originally I had implemented this with a constant for the number of
bits. But since we don't support an arbitrary number of bits for the
length, this didn't really make sense.

Now we're just taking a type parameter for the length, which is
restricted to the allowed types by a trait bound.