Deserializer supports peek mode
In this mode, any* deserialized values are peeked instead of read (the
buffer cursor is not advanced).

This will be used internally to support deserializing the untagged enum
used within the ListTerm type, and probably should not be used for other
cases. Ultimately this is a hack to avoid writing two
serde::Deserializer implementations.

* The string types that we support are not implemented in this mode, as
we do not support them as the terminator type in ListTerm.