Deserializer has limited enum support
This implementation is very specific to how the ListTerm type works,
because it uses what are effectively untagged enums.

Untagged enums are only supported in serde with formats that are
self-describing (UO network protocol is not). This is because there is
no generic way to distinguish variants from the serialized data.

In our case, we know that the variants in a ListTerm element have no
overlapping bit patterns, because one variant is a terminator which must
be null, and the other is a value which cannot have a null value for the
first N bytes (determined by the terminator type).

In theory this implementation can support any terminator sentinel value,
as long as the bit pattern for it is not valid in the first N bytes of a
value in the list. However, the implementation of TerminatorVariant
effectively limits the shape of supported enums to matching that of the
ListTerm element type.