ListTerm type can be deserialized
This is a significant rewrite, including switching to use a type
parameter for the terminator type rather than a constant parameter for
the number of bits (same as with List).

ListTerm is now represented in the serde data model as a sequence of
enums with two variants: value and terminator. The biggest change to
allow for this new approach is with the deserialization code. Both the
list type itself and the element enum type require their own custom
serde::Deserialize implementation and custom Visitor type.