Fixed-length packets have explicit size attr
It's not reasonable to determine the serialized size of a type using
serde's derived Serialize implementation, without having a constructed
version of that type to process.

However, it can be helpful to determine whether the deserialization
process has overrun or underrun the serialized data. For packets which
do not include a size field, we need to know this information at compile
time.

Rather than require each fixed-length packet to implement Default so
that we can use the serialization-based approach to determining its
size, the size should be manually calculated and provided as part of the
packet macro attributes.

This splits the kind of packet definition into three types:
* Fixed - with a required id and size
* Var - with a required id (runtime size field is included)
* Extended - with a required extended id (runtime size field is included)