Decoder checks readable amount before deserialize
Because the deserialization process reads from the BytesMut, we need to
verify that it has sufficient data to deserialize the complete packet.

This fixes a bug where the BytesMut may have had sufficient data to
check the packet header, but would fail partway through deserialization.
Previously we considered such cases to be errors, but we should have
been returning Ok(None) to indicate that more data is required before
decoding should be attempted again.