Server shutdown bool is internal
This simplifies the server's shutdown logic to work with an AtomicBool
that is created as a private field on Server, rather than received (in
an Arc) from whichever thread constructs the Server.

Because the Server itself is wrapped in an Arc when constructed by the
game binary, we can just call a method on it which will atomically store
the shutdown signal.