Replace custom uop hash with lookup3 hash function
It turns out I should have done a bit of research on the hash function
code before translating it from the C# implementation in the ClassicUO
project.

As indicated by
https://github.com/cbnolok/LegacyMULConverter-N/blob/2faebcec27263e323613af3b4b9bd83a80584105/LegacyMUL/Converter.cs#L54
the hash function used for UOP file paths is actually lookup3 by Bob
Jenkins. This replaces the custom implementation with one from the
hashers crate.