UOPackageFile has semantic fields, not file data
This is partly due to file entries having different data in version 4
and 5 of a UOP package, and partly because we do not need to represent
this data when working with the files in memory.

This also allows us to more easily create a set of files with their
semantic data and then write them to a UOP file without having to worry
about the specifics of the data format.