Encoder impls are more flexible
Specifically, they do not require that the encoded item type is
convertable to a Packet parameterized on that same item type. Instead,
we now use the associated type Content from the IntoPacket trait.

This allows us to send extended packet types, which convert from a
structure T into a Packet<(u16, T)>.