Generate more compact packages
We still try to align block headers and file entries starting on page
boundaries, but this reduces some of the wasted space from that.

If we want to support packages with many small files, doing so compactly
would require more work to sort files by size and group them in without
requiring page alignment.